requirements.txt
# Stripe MRR Dashboard - Python Dependencies
# ==========================================

# Stripe API
stripe==14.3.0

# Google Cloud BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.0.0

# Flask API Server
flask>=2.0.0
flask-cors>=3.0.0

# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Visualization (optional - for static charts)
matplotlib>=3.7.0
